+++ indirect_gwas/src/utils.hpp
@@ -17,7 +17,7 @@ struct ColumnSpec
 // Create a struct to store a matrix and its row and column names
 struct LabeledMatrix
 {
-    Eigen::MatrixXd data;
+    Eigen::MatrixXf data;
     std::vector<std::string> row_names;
     std::vector<std::string> column_names;
 };
@@ -25,20 +25,20 @@ struct LabeledMatrix
 // Create a struct to store a vector and its names
 struct LabeledVector
 {
-    Eigen::VectorXd data;
+    Eigen::VectorXf data;
     std::vector<std::string> names;
 };
 
 struct ResultsChunk
 {
     std::vector<std::string> variant_ids;
-    Eigen::MatrixXd beta;
-    Eigen::MatrixXd std_error;
-    Eigen::MatrixXd t_statistic;
-    Eigen::MatrixXd neg_log10_p_value;
-    Eigen::VectorXd sample_size;
+    Eigen::MatrixXf beta;
+    Eigen::MatrixXf std_error;
+    Eigen::MatrixXf t_statistic;
+    Eigen::MatrixXf neg_log10_p_value;
+    Eigen::VectorXf sample_size;
 };
 
-double compute_log_p_value(double t_statistic, unsigned int degrees_of_freedom);
+float compute_log_p_value(float t_statistic, unsigned int degrees_of_freedom);
 
 std::string get_formatted_time();
